requirements.txt
Flask
dash
dash_core_components
dash_html_components
pandas_datareader
gunicorn==19.0.0
dash_table
dash_bootstrap_components
geopandas